requirements.txt
numpy
google-genai
dotenv